test: run tests with lightningcss and postcss
