+++ playground/css-sourcemap/vite.config-lightningcss.js
@@ -1,0 +1,11 @@
+import { defineConfig, mergeConfig } from 'vite'
+import baseConfig from './vite.config.js'
+
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    css: {
+      transformer: 'lightningcss',
+    },
+  }),
+)
