+++ playground/css/__tests__/lightningcss/lightningcss.spec.ts
@@ -1,0 +1,4 @@
+// NOTE: a separate directory from `playground/css` is created by playground/vitestGlobalSetup.ts
+import { tests } from '../tests'
+
+tests(true)
