+++ playground/vitestGlobalSetup.ts
@@ -43,7 +43,7 @@ export async function setup({ provide }: TestProject): Promise<void> {
     })
   // also setup dedicated copy for "variant" tests
   for (const [original, variants] of [
-    ['css', ['sass-legacy', 'sass-modern']],
+    ['css', ['sass-legacy', 'sass-modern', 'lightningcss']],
     ['css-sourcemap', ['sass-legacy', 'sass-modern']],
   ] as const) {
     for (const variant of variants) {
